feat: 'volt profile destroy' now takes multiple arguments (fix #211)

+++ cmd/profile.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/hashicorp/go-multierror"
 	"github.com/vim-volt/volt/lockjson"
 	"github.com/vim-volt/volt/logger"
 	"github.com/vim-volt/volt/pathutil"
@@ -306,23 +307,11 @@ func (cmd *profileCmd) doDestroy(args []string) error {
 		logger.Error("'volt profile destroy' receives profile name.")
 		return nil
 	}
-	profileName := args[0]
 
 	// Read lock.json
 	lockJSON, err := lockjson.Read()
 	if err != nil {
 		return errors.New("failed to read lock.json: " + err.Error())
-	}
-
-	// Return error if current profile matches profileName
-	if lockJSON.CurrentProfileName == profileName {
-		return errors.New("cannot destroy current profile: " + profileName)
-	}
-
-	// Return error if profiles[]/name does not match profileName
-	index := lockJSON.Profiles.FindIndexByName(profileName)
-	if index < 0 {
-		return errors.New("profile '" + profileName + "' does not exist")
 	}
 
 	// Begin transaction
@@ -332,14 +321,33 @@ func (cmd *profileCmd) doDestroy(args []string) error {
 	}
 	defer transaction.Remove()
 
-	// Remove the specified profile
-	lockJSON.Profiles = append(lockJSON.Profiles[:index], lockJSON.Profiles[index+1:]...)
+	var merr *multierror.Error
+	for i := range args {
+		profileName := args[i]
 
-	// Remove $VOLTPATH/rc/{profile} dir
-	rcDir := pathutil.RCDir(profileName)
-	os.RemoveAll(rcDir)
-	if pathutil.Exists(rcDir) {
-		return errors.New("failed to remove " + rcDir)
+		// Skip if current profile matches profileName
+		if lockJSON.CurrentProfileName == profileName {
+			merr = multierror.Append(merr, errors.New("cannot destroy current profile: "+profileName))
+			continue
+		}
+		// Skip if profiles[]/name does not match profileName
+		index := lockJSON.Profiles.FindIndexByName(profileName)
+		if index < 0 {
+			merr = multierror.Append(merr, errors.New("profile '"+profileName+"' does not exist"))
+			continue
+		}
+
+		// Remove the specified profile
+		lockJSON.Profiles = append(lockJSON.Profiles[:index], lockJSON.Profiles[index+1:]...)
+
+		// Remove $VOLTPATH/rc/{profile} dir
+		rcDir := pathutil.RCDir(profileName)
+		os.RemoveAll(rcDir)
+		if pathutil.Exists(rcDir) {
+			return errors.New("failed to remove " + rcDir)
+		}
+
+		logger.Info("Deleted profile '" + profileName + "'")
 	}
 
 	// Write to lock.json
@@ -348,9 +356,7 @@ func (cmd *profileCmd) doDestroy(args []string) error {
 		return err
 	}
 
-	logger.Info("Deleted profile '" + profileName + "'")
-
-	return nil
+	return merr.ErrorOrNil()
 }
 
 func (cmd *profileCmd) doRename(args []string) error {
